Added notification class skeleton, renamed NotificationHelper to NotificationSender
