+++ app/src/main/java/io/github/tstewart/todayi/notifications/DailyAlarmReceiver.java
@@ -1,0 +1,14 @@
+package io.github.tstewart.todayi.notifications;
+
+import android.content.BroadcastReceiver;
+import android.content.Context;
+import android.content.Intent;
+
+/* Receives Alarm requests for Daily Reminder notifications, and sends a request to NotificationSender */
+public class DailyAlarmReceiver extends BroadcastReceiver {
+
+    @Override
+    public void onReceive(Context context, Intent intent) {
+
+    }
+}
